chore(vue2): Fix build process
- Upgrade babel dependencies
- Use rollup directly instead of building via gulp
- Split rollup config into UMD and ESM
- Stop transpiling ESM to ES5; just transpile Vue files to JS
- Stop building helper JSON in build step (it's not published)
