+++ bindings/vue2/vue-onsenui/package.json
@@ -4,10 +4,12 @@
   "description": "Vue.js bindings for Onsen UI",
   "author": "Onsen UI Team <team@monaca.io>",
   "scripts": {
-    "build": "cross-env NODE_ENV=production gulp build && npm run build:minify && npm run build:helper-json",
-    "build:helper-json": "(cd ../../ && npm run build:docs) && gulp build:helper-json",
+    "build": "npm run build:esm && npm run build:umd && npm run build:minify",
+    "build:esm": "cross-env NODE_ENV=production rollup -c rollup.config.esm.js && cpx 'src/**/!(index.umd).js' esm",
     "build:minify": "uglifyjs dist/vue-onsenui.js -c -m --comments '/vue-onsenui/' --output dist/vue-onsenui.min.js",
+    "build:umd": "cross-env NODE_ENV=production rollup -c rollup.config.umd.js",
     "clean": "rimraf ./esm ./dist",
+    "helper-json": "(cd ../../ && npm run build:docs) && gulp build:helper-json",
     "prepublish": "npm run build",
     "test": "gulp test"
   },
@@ -42,17 +44,30 @@
     "onsenui"
   ],
   "homepage": "https://onsen.io/",
+  "babel": {
+    "presets": [
+      [
+        "@babel/preset-env",
+        {
+          "modules": false
+        }
+      ]
+    ]
+  },
   "peerDependencies": {
     "onsenui": "~2.11.0",
     "vue": "~2.6.0"
   },
   "devDependencies": {
-    "babel-core": "^6.26.0",
+    "@babel/core": "^7.17.2",
+    "@babel/plugin-external-helpers": "^7.16.7",
+    "@babel/preset-env": "^7.16.11",
+    "@rollup/plugin-babel": "^5.3.0",
     "babel-eslint": "^8.0.1",
-    "babel-plugin-external-helpers": "^6.22.0",
     "babel-plugin-transform-runtime": "^6.23.0",
     "babel-preset-env": "^1.4.0",
     "babel-preset-stage-3": "^6.22.0",
+    "cpx": "^1.5.0",
     "cross-env": "^1.0.8",
     "eslint": "^6.7.2",
     "eslint-plugin-vue": "^6.2.2",
@@ -68,16 +83,10 @@
     "merge-stream": "^1.0.1",
     "rimraf": "^3.0.2",
     "rollup": "^2.52.3",
-    "rollup-plugin-alias": "^1.4.0",
-    "rollup-plugin-babel": "^3.0.2",
-    "rollup-plugin-commonjs": "^8.2.6",
     "rollup-plugin-eslint": "^4.0.0",
-    "rollup-plugin-execute": "^1.0.0",
-    "rollup-plugin-filesize": "^1.4.2",
     "rollup-plugin-node-resolve": "^3.0.0",
     "rollup-plugin-progress": "^1.1.2",
     "rollup-plugin-replace": "^2.0.0",
-    "rollup-plugin-visualizer": "^0.3.1",
     "rollup-plugin-vue": "4.2.0",
     "uglify-js": "^3.1.7",
     "vue": "^2.5.2",
