--- bindings/vue2/vue-onsenui/rollup.config.js
@@ -1,122 +1,0 @@
-import path from 'path';
-import pkg from './package.json';
-import corePkg from '../../../package.json';
-import dateformat from 'dateformat';
-
-// Rollup plugins
-import babel from 'rollup-plugin-babel';
-import { eslint } from 'rollup-plugin-eslint';
-import resolve from 'rollup-plugin-node-resolve';
-import commonjs from 'rollup-plugin-commonjs';
-import replace from 'rollup-plugin-replace';
-import filesize from 'rollup-plugin-filesize';
-import progress from 'rollup-plugin-progress';
-import visualizer from 'rollup-plugin-visualizer';
-import alias from 'rollup-plugin-alias';
-import vue from 'rollup-plugin-vue';
-
-const local = (...args) => path.resolve(__dirname, ...args);
-
-const babelrc = Object.assign({}, corePkg.babel);
-babelrc.babelrc = babelrc.presets[0][1].modules = false;
-babelrc.plugins = ['external-helpers'];
-babelrc.exclude = [local('node_modules/**'), local('../../build/**')];
-
-const globals = { 'onsenui': 'ons', 'onsenui/esm': 'ons' },
-  external = id => /^onsenui/.test(id),
-  banner = `/* ${pkg.name} v${pkg.version} - ${dateformat(new Date(), 'yyyy-mm-dd')} */\n`;
-
-const builds = [
-  // Vue bindings UMD
-  {
-    input: 'src/index.umd.js',
-    external,
-    output: {
-      file: 'dist/vue-onsenui.js',
-      format: 'umd',
-      name: 'VueOnsen',
-      sourcemap: 'inline',
-      globals,
-      banner,
-    },
-    plugins: [
-      eslint({
-        exclude: [
-          /vue&type=/,
-          /node_modules/
-        ]
-      }),
-      resolve({ extensions: ['.js', '.vue'] }),
-      replace({
-        // Remove undefined imports to slightly reduce UMD bundle size
-        include: './src/components/*.vue',
-        'import \'onsenui/esm/elements/': '// \'',
-      }),
-      vue(),
-      babel(babelrc),
-      progress(),
-      filesize(),
-      visualizer({
-        filename: 'module-stats.umd.html',
-        sourcemap: true,
-      })
-    ],
-  },
-
-  // Vue bindings ES Modules
-  {
-    input: 'src/index.esm.js',
-    external,
-    output: {
-      file: 'esm/index.js',
-      format: 'es',
-      name: 'VueOnsenESM',
-      sourcemap: 'inline',
-      globals,
-      banner,
-    },
-    plugins: [
-      babel(babelrc),
-      progress(),
-      filesize(),
-      visualizer({
-        filename: 'module-stats.esm.html',
-        sourcemap: false,
-      }),
-    ],
-  },
-];
-
-// Make it work with rollup CLI and Gulp
-builds.devConfig = {
-  input: local('examples/main.js'),
-  output: {
-    file: local('examples/build.js'),
-    format: 'umd',
-    name: 'VueOnsenDev',
-    sourcemap: 'inline',
-  },
-  plugins: [
-    alias({
-      resolve: ['.js', '.vue', '\/index.js'],
-      vue: local('node_modules/vue/dist/vue.esm.js'),
-      'vue-onsenui/esm/components': local('src', 'components'),
-      'vue-onsenui/esm': local('src', 'index.esm.js'),
-      'vue-onsenui': local('src', 'index.umd.js'),
-      'onsenui/esm': local('../../build/esm'),
-      'onsenui': local('../../build/js/onsenui'),
-    }),
-    resolve({ extensions: ['.js', '.vue'] }),
-    replace({
-      'import \'onsenui/esm/elements/': '// \'',
-      'process.env.NODE_ENV': JSON.stringify( 'development' )
-    }),
-    commonjs({ include: ['node_modules/**', local('../../build/js/*')] }),
-    vue(),
-    babel(babelrc),
-    progress(),
-    filesize(),
-  ],
-};
-
-export default builds;
